resampling using doubles instead of floats getsamples to combine some overhead and return multiple samples in one call Add bisinc and bicubic

+++ beam-core/src/test/java/org/esa/beam/framework/dataop/resamp/BilinearInterpolationResamplingTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2010 Brockmann Consult GmbH (info@brockmann-consult.de)
+ * Copyright (C) 2014 Brockmann Consult GmbH (info@brockmann-consult.de)
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the Free
@@ -40,18 +40,18 @@ public class BilinearInterpolationResamplingTest extends TestCase {
     public void testIndexAndSample() throws Exception {
         final Resampling.Index index = resampling.createIndex();
 
-        testIndexAndSample(index, -.5f, 0.0f, 0, 0, 0, 0, 0.0f, 0.5f, 10.0f);
+        testIndexAndSample(index, -.5f, 0.0f, 0.0, 0.0, 0.0, 0.0, 0.0f, 0.5f, 10.0f);
 
-        testIndexAndSample(index, 0.0f, 0.0f, 0, 0, 0, 0, 0.5f, 0.5f, 10.0f);
-        testIndexAndSample(index, 0.5f, 0.0f, 0, 1, 0, 0, 0.0f, 0.5f, 10.0f);
-        testIndexAndSample(index, 1.0f, 0.0f, 0, 1, 0, 0, 0.5f, 0.5f, 15.0f);
-        testIndexAndSample(index, 1.5f, 0.0f, 1, 2, 0, 0, 0.0f, 0.5f, 20.0f);
-        testIndexAndSample(index, 2.0f, 0.0f, 1, 2, 0, 0, 0.5f, 0.5f, 25.0f);
-        testIndexAndSample(index, 2.5f, 0.0f, 2, 3, 0, 0, 0.0f, 0.5f, 30.0f);
-        testIndexAndSample(index, 3.0f, 0.0f, 2, 3, 0, 0, 0.5f, 0.5f, 35.0f);
-        testIndexAndSample(index, 3.5f, 0.0f, 3, 4, 0, 0, 0.0f, 0.5f, 40.0f);
-        testIndexAndSample(index, 4.0f, 0.0f, 3, 4, 0, 0, 0.5f, 0.5f, 45.0f);
-        testIndexAndSample(index, 4.5f, 0.0f, 4, 4, 0, 0, 0.0f, 0.5f, 50.0f);
+        testIndexAndSample(index, 0.0f, 0.0f, 0.0, 0.0, 0.0, 0.0, 0.5f, 0.5f, 10.0f);
+        testIndexAndSample(index, 0.5f, 0.0f, 0.0, 1.0, 0.0, 0.0, 0.0f, 0.5f, 10.0f);
+        testIndexAndSample(index, 1.0f, 0.0f, 0.0, 1.0, 0.0, 0.0, 0.5f, 0.5f, 15.0f);
+        testIndexAndSample(index, 1.5f, 0.0f, 1.0, 2.0, 0.0, 0.0, 0.0f, 0.5f, 20.0f);
+        testIndexAndSample(index, 2.0f, 0.0f, 1.0, 2.0, 0.0, 0.0, 0.5f, 0.5f, 25.0f);
+        testIndexAndSample(index, 2.5f, 0.0f, 2.0, 3.0, 0.0, 0.0, 0.0f, 0.5f, 30.0f);
+        testIndexAndSample(index, 3.0f, 0.0f, 2.0, 3.0, 0.0, 0.0, 0.5f, 0.5f, 35.0f);
+        testIndexAndSample(index, 3.5f, 0.0f, 3.0, 4.0, 0.0, 0.0, 0.0f, 0.5f, 40.0f);
+        testIndexAndSample(index, 4.0f, 0.0f, 3.0, 4.0, 0.0, 0.0, 0.5f, 0.5f, 45.0f);
+        testIndexAndSample(index, 4.5f, 0.0f, 4.0, 4.0, 0.0, 0.0, 0.0f, 0.5f, 50.0f);
 
         testIndexAndSample(index, 5.0f, 0.0f, 4, 4, 0, 0, 0.5f, 0.5f, 50.0f);
         testIndexAndSample(index, 5.5f, 0.0f, 4, 4, 0, 0, 0.0f, 0.5f, 50.0f);
@@ -60,12 +60,12 @@ public class BilinearInterpolationResamplingTest extends TestCase {
     }
 
     private void testIndexAndSample(final Resampling.Index index,
-                      float x, float y,
-                      int i1Exp, int i2Exp,
-                      int j1Exp, int j2Exp,
-                      float kiExp,
-                      float kjExp,
-                      float sampleExp) throws Exception {
+                                    float x, float y,
+                                    double i1Exp, double i2Exp,
+                                    double j1Exp, double j2Exp,
+                                    float kiExp,
+                                    float kjExp,
+                                    float sampleExp) throws Exception {
         resampling.computeIndex(x, y, raster.getWidth(), raster.getHeight(), index);
         assertEquals(i1Exp, index.i[0]);
         assertEquals(i2Exp, index.i[1]);
@@ -73,7 +73,7 @@ public class BilinearInterpolationResamplingTest extends TestCase {
         assertEquals(j2Exp, index.j[1]);
         assertEquals(kiExp, index.ki[0], 1e-5f);
         assertEquals(kjExp, index.kj[0], 1e-5f);
-        float sample = resampling.resample(raster, index);
+        double sample = resampling.resample(raster, index);
         assertEquals(sampleExp, sample, 1e-5f);
     }
 }
